modularize dialects into their own submodules
remove assertion in copy sql, conversion to_string is unfinished
